first crack at it, on just graphql side
