+++ crates/sui-graphql-rpc/Cargo.toml
@@ -70,3 +70,4 @@ tower.workspace = true
 
 [features]
 pg_integration = []
+pg_backend = []
