+++ crates/sui-graphql-rpc/src/context_data/mod.rs
@@ -1,8 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod db_backend;
 pub(crate) mod db_data_provider;
 pub mod db_query_cost;
 pub(crate) mod package_cache;
+#[cfg(feature = "pg_backend")]
+pub(crate) mod pg_backend;
 
 pub const DEFAULT_PAGE_SIZE: u64 = 10;
